Add SMS processing through Twilio.
Tried to migrate this branch from another repo, but could only do
that for one branch, not two. Therefore, losing the commit history
by adding files manually, but all work here. Commit history was
short (and uninformative) to begin with, so this should be fine.

Code, tests, requirements, and README all up-to-date.