Add multiple capabilities around SMS delivery.
This PR addresses a number of areas around the automated delivery
of SMS messages to NGS2 participants. They are:
* Add ability to send messages to the Philippines. Now that the
sample for work includes the Philippines, additional functionality
to process numbers and send messages has been created.
* Gather bitlinks by using the Bitly API. For some messages, we
need to send web links to the empanelment survey (or experiment
instance) and those addresses are generally too long. This PR adds
capabilities to process individualized weblinks through Bitly's
API and return bitlinks that can be inserted into SMS messages.
* Modify how messaged text is processed. With the introduction of
individualized links to add to what has otherwise been static
content to all recipients, there needed to be modifications for
when links are present.